Обновляет тесты
