+++ src/components/genres-list/genres-list.test.js
@@ -1,16 +1,28 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import GenresList from "./genres-list.jsx";
+import {GenresList} from "./genres-list.jsx";
+import {Provider} from "react-redux";
+import configureStore from "redux-mock-store";
 
-const genresList = [`All genres`, `Drama`, `Comedy`, `Thriller`];
+const mockStore = configureStore([]);
+
+const genres = [`All genres`, `Drama`, `Comedy`, `Thriller`];
 
 it(`Genre should render correctly`, () => {
+  const store = mockStore({
+    activeGenre: genres[0],
+    genresList: genres,
+  });
+
   const tree = renderer
-     .create(<GenresList
-       onGenreClick={() => {}}
-       activeGenre={genresList[0]}
-       genresList={genresList}
-     />)
+     .create(
+         <Provider store={store}>
+           <GenresList
+             onGenreClick={() => {}}
+             activeGenre={genres[0]}
+             genresList={genres}
+           />
+         </Provider>)
      .toJSON();
 
   expect(tree).toMatchSnapshot();
