+++ src/components/more-like-this/more-like-this.test.js
@@ -1,16 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import {App} from "./app.jsx";
-import {Provider} from "react-redux";
-import configureStore from "redux-mock-store";
-
-const mockStore = configureStore([]);
-
-const PROMO_MOVIE = {
-  title: `Harry Potter`,
-  genre: `Comedy`,
-  releaseDate: 2009,
-};
+import MoreLikeThis from "./more-like-this.jsx";
 
 const movies = [
   {
@@ -29,7 +19,6 @@ const movies = [
     },
     director: `David Yates`,
     actors: `Eddie Redmayne, Katherine Waterston, Dan Fogler, Alison Sudol, Ezra Miller, Zoë Kravitz`,
-    runtime: `1h 39m`,
   },
   {
     previewImage: `img/bohemian-rhapsody.jpg`,
@@ -47,45 +36,17 @@ const movies = [
     },
     director: `Bryan Singer`,
     actors: `Rami Malek, Lucy Boynton, Gwilym Lee, Ben Hardy, Joe Mazzello, Aidan Gillen, Tom Hollander`,
-    runtime: `1h 39m`,
   },
 ];
 
-const reviews = [
-  {
-    id: 1,
-    text: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
-    author: `Kate Muir`,
-    date: `December 24, 2016`,
-    rating: 8.9,
-  },
-  {
-    id: 2,
-    text: `Anderson's films are too precious for some, but for those of us willing to lose ourselves in them, they're a delight. "The Grand Budapest Hotel" is no different, except that he has added a hint of gravitas to the mix, improving the recipe.`,
-    author: `Bill Goodykoontz`,
-    date: `November 18, 2015`,
-    rating: 8.0,
-  },
-];
-
-it(`App should render correctly`, () => {
-  const store = mockStore({
-    promoMovie: PROMO_MOVIE,
-    reviews,
-    currentMovieCard: movies[0],
-    movies,
-  });
-
+it(`MoviesList should render correctly`, () => {
   const tree = renderer
      .create(
-         <Provider store={store}>
-           <App
-             promoMovie={PROMO_MOVIE}
-             reviews={reviews}
-             currentMovieCard={movies[0]}
-             handleMovieCardClick={() => {}}
-           />
-         </Provider>, {
+         <MoreLikeThis
+           movies={movies}
+           onMovieCardClick={() => {}}
+           onActiveCardChange={() => {}}
+         />, {
            createNodeMock: () => {
              return {};
            }
