+++ src/components/movie-page/movie-page.test.js
@@ -1,6 +1,10 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import MoviePage from "./movie-page.jsx";
+import {MoviePage} from "./movie-page.jsx";
+import {Provider} from "react-redux";
+import configureStore from "redux-mock-store";
+
+const mockStore = configureStore([]);
 
 const movies = [
   {
@@ -77,15 +81,20 @@ const reviews = [
 ];
 
 it(`MoviePage should render correctly`, () => {
+  const store = mockStore({
+    movies,
+  });
 
   const tree = renderer
      .create(
-         <MoviePage
-           movie={movies[0]}
-           movies={movies}
-           reviews={reviews}
-           onMovieCardClick={() => {}}
-         />, {
+         <Provider store={store}>
+           <MoviePage
+             movie={movies[0]}
+             movies={movies}
+             reviews={reviews}
+             onMovieCardClick={() => {}}
+           />
+         </Provider>, {
            createNodeMock: () => {
              return {};
            }
