+++ src/components/movies-list/movies-list.test.js
@@ -1,6 +1,10 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import MoviesList from "./movies-list.jsx";
+import {MoviesList} from "./movies-list.jsx";
+import {Provider} from "react-redux";
+import configureStore from "redux-mock-store";
+
+const mockStore = configureStore([]);
 
 const movies = [
   {
@@ -40,13 +44,22 @@ const movies = [
 ];
 
 it(`MoviesList should render correctly`, () => {
+  const store = mockStore({
+    movies,
+    shownMovieCards: 8,
+  });
+
   const tree = renderer
      .create(
-         <MoviesList
-           movies={movies}
-           onMovieCardClick={() => {}}
-           onActiveCardChange={() => {}}
-         />, {
+         <Provider store={store}>
+           <MoviesList
+             movies={movies}
+             onMovieCardClick={() => {}}
+             onActiveCardChange={() => {}}
+             handleShowMoreButtonClick={() => {}}
+             shownMovieCards={8}
+           />
+         </Provider>, {
            createNodeMock: () => {
              return {};
            }
